Fix pgstreamrep_monitor() and status sequence.
Formerly the RA used a sequence of pgstreamrep_set_status_variables()
and pgstreamrep_get_status() to get the current state of the cluster.
However, this doesn't work well in case an PostgreSQL instance is not able
to start (or due to constraints not allowed on a specific node). Change this
by switching the sequence to pgstreamrep_get_status() and
pgstreamrep_set_status_variables(). This checks the PID file and status
of the whole cluster *before* extracting the cluster role.

While at it, fix some minor identation quirks.