Further fixes following commit 67603c304089ca9404e2804bebe3d5dd9567f7ba
It turns out, that the mentioned fixes for the monitoring
sequence in pgstreamrep_monitor() is not enough; testing shows,
that in case *no* cluster resource is running, the agent falls through
the routine and claims, that the unconfigured role is okay, which is
obviously wrong. Instead, return the exact error code in case
pgstreamrep_get_status() fails, forcing the RA into failure condition.