Implement node specific attributes for role and XLOG position.
This teaches the pgstreamrep RA to set node attributes to reflect
current state of the RA on a specific node. The current node attributes
set by the RA are

pgstreamrep-role: Current role of the PostgreSQL instance
pgstreamrep-xlogpos: Current position in the XLOG

This allows arbitrary actions on certain conditions triggered by the RA.
One example are location constraints, where specific resources are bound
to the role of the running PostgreSQL instance, for example an IP address
that should stay with a slave node. This could now be accomplished by such
a rule:

location locStandbyIP pgStandbyIP \
	rule $id="run-standby-ip" inf: pgstreamrep-role eq Slave

Other workflows are possible and the feature is likely to be improved in
the near future ;)